0.0.11

+++ Sources/CGGeometry/CGFloat.swift
@@ -24,7 +24,7 @@ import UIKit
 
 public extension CGFloat {
     
-    static var max = Float.greatestFiniteMagnitude
+    static var max = CGFloat.greatestFiniteMagnitude
     
 }
